+++ MoltenVK/MoltenVK/GPUObjects/MVKQueue.mm
@@ -158,7 +158,7 @@ MVKQueue::MVKQueue(MVKDevice* device, MVKQueueFamily* queueFamily, uint32_t inde
 	initName();
 	initExecQueue();
 	initMTLCommandQueue();
-	initGPUCaptureScopes();
+	initGPUCaptureScopes();		// After initMTLCommandQueue()
 }
 
 void MVKQueue::initName() {
@@ -199,6 +199,8 @@ void MVKQueue::initGPUCaptureScopes() {
 		_submissionCaptureScope->makeDefault();
 	}
 	_submissionCaptureScope->beginScope();	// Allow Xcode to capture the first frame if desired.
+
+	getInstance()->startAutoGPUCapture(MVK_CONFIG_AUTO_GPU_CAPTURE_SCOPE_FRAME, _mtlQueue);
 }
 
 MVKQueue::~MVKQueue() {
@@ -379,6 +381,7 @@ void MVKQueuePresentSurfaceSubmission::execute() {
 	auto cs = _queue->_submissionCaptureScope;
 	cs->endScope();
 	cs->beginScope();
+	stopAutoGPUCapture();
 
 	this->destroy();
 }
@@ -388,6 +391,16 @@ id<MTLCommandBuffer> MVKQueuePresentSurfaceSubmission::getMTLCommandBuffer() {
 	setLabelIfNotNil(mtlCmdBuff, @"vkQueuePresentKHR CommandBuffer");
 	[mtlCmdBuff enqueue];
 	return mtlCmdBuff;
+}
+
+
+void MVKQueuePresentSurfaceSubmission::stopAutoGPUCapture() {
+	MVKInstance* mvkInst = _queue->getInstance();
+	const MVKConfiguration* pMVKConfig = mvkInst->getMoltenVKConfiguration();
+	if (_queue->_queueFamily->getIndex() == pMVKConfig->defaultGPUCaptureScopeQueueFamilyIndex &&
+		_queue->_index == pMVKConfig->defaultGPUCaptureScopeQueueIndex) {
+		mvkInst->stopAutoGPUCapture(MVK_CONFIG_AUTO_GPU_CAPTURE_SCOPE_FRAME);
+	}
 }
 
 MVKQueuePresentSurfaceSubmission::MVKQueuePresentSurfaceSubmission(MVKQueue* queue,